Created Backend with express framework and API endpoints
Created Function:
ConnectDB: To connect to MySQL Database hosted on local machine

setAppOptions: To set JSON body parser and CORS option for express app

POST /shortUrls: POST endpoint to receive Full Url link and convert to short Url. Checks if Url is vaild or not, sends appropriate json response back.

GET checkUrlShortCount: Gets url to be checked in the database and returns the count for the number of Full Url found else if no Url are found sends 0.